requirements.txt
Flask
Pillow
exifread
